+++ R/readConfiguration.R
@@ -580,7 +580,7 @@ checkScenario <- function(scenario = defaultScenario())
   if (is.null.or.empty(scenario$initConfigurations)) {
     scenario$initConfigurations <- NULL
   } else if (!is.data.frame(scenario$initConfigurations) && !is.matrix(scenario$initConfigurations)) {
-    irace.error("if given, initConfigurations must be a matrix or data.frame")
+    irace.error("if given, 'initConfigurations' must be a matrix or data.frame.")
   }
   
   # We have characters everywhere, set to the right types to avoid
@@ -595,8 +595,8 @@ checkScenario <- function(scenario = defaultScenario())
     if (is.na(scenario[[param]]))
       next # Allow NA default values
     p <- suppressWarnings(as.numeric(p))
-    if (is.null(p) || is.na (p) || !is.wholenumber(p))
-      irace.error (quote.param (param), " must be an integer.")
+    if (is.null(p) || is.na (p) || !is.wholenumber(p) || p < 0)
+      irace.error (quote.param (param), " must be a non-negative integer.")
     scenario[[param]] <- as.integer(p)
   }
 
@@ -815,6 +815,7 @@ printScenario <- function(scenario)
 #'      \item{`targetRunnerLauncher`}{Executable that will be used to launch the target runner, when \code{targetRunner} cannot be executed directly (e.g., a Python script in Windows). (Default: `""`)}
 #'      \item{`targetCmdline`}{Command-line arguments provided to \code{targetRunner} (or \code{targetRunnerLauncher} if defined). The substrings \code{\{configurationID\}}, \code{\{instanceID\}},  \code{\{seed\}},  \code{\{instance\}}, and \code{\{bound\}} will be replaced by their corresponding values. The substring \code{\{targetRunnerArgs\}} will be replaced by the concatenation of the switch and value of all active parameters of the particular configuration being evaluated.  The substring \code{\{targetRunner\}}, if present, will be replaced by the value of \code{targetRunner} (useful when using \code{targetRunnerLauncher}). (Default: `"{configurationID} {instanceID} {seed} {instance} {bound} {targetRunnerArgs}"`)}
 #'      \item{`targetRunnerRetries`}{Number of times to retry a call to \code{targetRunner} if the call failed. (Default: `0`)}
+#'      \item{`targetRunnerTimeout`}{Timeout in seconds of any \code{targetRunner} call (only applies to \code{target-runner} executables not to R functions), ignored if 0. (Default: `0`)}
 #'      \item{`targetRunnerData`}{Optional data passed to \code{targetRunner}. This is ignored by the default \code{targetRunner} function, but it may be used by custom \code{targetRunner} functions to pass persistent data around. (Default: `""`)}
 #'      \item{`targetRunnerParallel`}{Optional R function to provide custom parallelization of \code{targetRunner}. (Default: `""`)}
 #'      \item{`targetEvaluator`}{Optional script or R function that provides a numeric value for each configuration. See templates/target-evaluator.tmpl (Default: `""`)}